chore: project와 git repository를 분리함
