--- .settings/org.eclipse.wst.common.project.facet.core.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<faceted-project>
-  <fixed facet="jst.web"/>
-  <fixed facet="jst.java"/>
-  <installed facet="jst.web" version="2.5"/>
-  <installed facet="jst.java" version="1.6"/>
-</faceted-project>
